hugging/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
pydantic
langchain
google-generativeai
langchain-google-genai
numpy
faiss-cpu
sentence-transformers